+++ server/src/api/mod.rs
@@ -1,10 +1,9 @@
 //! API
 //!
 use crate::internal::*;
-use petshop_proto::api::User;
 use std::fmt;
 use tokio::sync::broadcast;
-use tonic::{Code, Request, Status};
+use tonic::{Code, Status};
 
 mod example;
 mod petshop;
@@ -13,10 +12,11 @@ mod tfb;
 /// API Server
 #[derive(Clone)]
 pub struct Api {
-    pub metrics: Arc<Metrics>,
-    pub csrf: Arc<Csrf>,
-    pub postgres: Arc<PostgresPool>,
     pub shutdown: Arc<broadcast::Sender<bool>>,
+    pub metrics: Arc<Metrics>,
+    pub postgres: Arc<PostgresPool>,
+    pub auth: Arc<Auth>,
+    pub csrf: Arc<Csrf>,
 
     /// This is only here for TFB fortunes endpoint
     pub tfb_handlebars: Arc<handlebars::Handlebars<'static>>,
@@ -40,7 +40,12 @@ impl Api {
         config: &Config,
         shutdown_tx: broadcast::Sender<bool>,
     ) -> Result<Self, XErr> {
+        let shutdown = Arc::new(shutdown_tx);
         let metrics = Arc::new(Metrics::from_config(config));
+        let postgres = Arc::new(PostgresPool::from_config(config, metrics.clone())?);
+
+        let auth = Arc::new(Auth::from_config(config, postgres.clone()));
+        let csrf = Arc::new(Csrf::from_config(config, metrics.clone()));
 
         let mut tfb_handlebars = handlebars::Handlebars::new();
         tfb_handlebars
@@ -48,12 +53,20 @@ impl Api {
             .expect("register template failed");
 
         Ok(Self {
-            metrics: metrics.clone(),
-            csrf: Arc::new(Csrf::from_config(config, metrics.clone())),
-            postgres: Arc::new(PostgresPool::from_config(config, metrics)?),
-            shutdown: Arc::new(shutdown_tx),
+            shutdown,
+            metrics,
+            postgres,
+            auth,
+            csrf,
             tfb_handlebars: Arc::new(tfb_handlebars),
         })
+    }
+
+    /// Sends shutdown signal to stop application
+    ///
+    /// This lets the application trigger a graceful exit rather than panicking
+    pub fn _shutdown(&self) {
+        self.shutdown.send(true).expect("shutdown failed");
     }
 
     pub fn metrics(&self) -> Arc<Metrics> {
@@ -64,17 +77,6 @@ impl Api {
         self.csrf.clone()
     }
 
-    pub fn postgres(&self) -> Arc<PostgresPool> {
-        self.postgres.clone()
-    }
-
-    /// Sends shutdown signal to stop application
-    ///
-    /// This lets the application trigger a graceful exit rather than panicking
-    pub fn shutdown(&self) {
-        self.shutdown.send(true).expect("shutdown failed");
-    }
-
     /// Returns an error if requests can not be served
     ///
     /// [More information on liveness/readiness probes](https://blog.colinbreck.com/kubernetes-liveness-and-readiness-probes-how-to-avoid-shooting-yourself-in-the-foot/)
@@ -83,49 +85,6 @@ impl Api {
         self.metrics.api_ready(postgres_ready.is_ok());
         postgres_ready?;
         Ok(())
-    }
-
-    /// Parses request metadata to extract authenticated user (works with auth example)
-    fn user_required(&self, request: &Request<()>) -> Result<User, Status> {
-        let email = request.metadata().get("x-auth-request-email");
-        let user = request.metadata().get("x-auth-request-user");
-        match (email, user) {
-            (Some(email), Some(user)) => match (email.to_str(), user.to_str()) {
-                (Ok(email), Ok(user)) => Ok(User {
-                    email: email.to_string(),
-                    name: user.to_string(),
-                }),
-                _ => Err(Status::unauthenticated(ERROR_USER_AUTHENTICATION)),
-            },
-            _ => Err(Status::unauthenticated(ERROR_USER_AUTHENTICATION)),
-        }
-    }
-
-    /// Parses request metadata to return authenticated user, which may be provided by oauth2-proxy
-    /// headers or by the authorization header
-    ///
-    /// FIXME: This is a placeholder function that accepts any authorization header as valid to demonstrate
-    /// supporting authentication via oauth2-proxy (for users) or via a header (for computers)
-    ///
-    /// In the auth example, this is made functional by adding an envoy listener that does not use the
-    /// ext_authz filter, so requests are still passed upstream where they can be checked by this function
-    ///
-    /// This example assumes that the application is going to manage/verify its own API keys and that
-    /// all private endpoints will call this function
-    ///
-    /// <https://cheatsheetseries.owasp.org/cheatsheets/REST_Security_Cheat_Sheet.html#api-keys>
-    fn auth_required(&self, request: &Request<()>) -> Result<User, Status> {
-        let auth = request.metadata().get("authorization");
-        match auth {
-            Some(auth) => match auth.to_str() {
-                Ok(auth) => Ok(User {
-                    email: "apiconsumer@petshop.com".to_string(),
-                    name: auth.to_string(),
-                }),
-                _ => self.user_required(request),
-            },
-            _ => self.user_required(request),
-        }
     }
 
     /// Validates request using derived validate method, logs validation errors