+++ server/src/services/mod.rs
@@ -1,6 +1,7 @@
 //! # Services
 //!
+mod auth;
 mod csrf;
 mod metrics;
 
-pub use crate::services::{csrf::*, metrics::*};
+pub use crate::services::{auth::*, csrf::*, metrics::*};